Refactor dupicated code that accesses basic attributes
several methods have dupicated code where only the attribute name is
different, but all other logic is the same.

Created a __get_obj_attr($attr) method.
This is then in turn called by the accessor method. This may not really
be needed but I like the idea.

$obj->season() calls $self->__get_obj_attr('season') to access the
attributes. $obj->month() does the same and so on where the logic is the
same.